[WIP] Browser: Add commands to open in horizontal / vertical split from command palette (#1690)
* Factor BrowserView to separate file

* Add browser commands from command palette
